# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Books API
  description: A simple API for managing books.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /books:
    get:
      summary: Get a list of books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              example:
                - id: 1
                  title: To Kill a Mockingbird
                  price: 12.99
                - id: 2
                  title: 1984
                  price: 10.99
      tags:
        - Books
    post:
      summary: Create a new book
      requestBody:
        description: Book data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the book
                price:
                  type: number
                  format: float
                  description: The price of the book
            example:
              title: The Catcher in the Rye
              price: 10.00
      responses:
        '201':
          description: The created book
          content:
            application/json:
              example:
                id: 11
                title: The Catcher in the Rye
                price: 10.00
      tags:
        - Books
  /books/{id}:
    get:
      summary: Get a book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the book
      responses:
        '200':
          description: A book
          content:
            application/json:
              example:
                id: 1
                title: To Kill a Mockingbird
                price: 12.99
        '404':
          description: Book not found
          content:
            application/json:
              example:
                error: Book not found
      tags:
        - Books
    delete:
      summary: Delete a book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the book to delete
      responses:
        '204':
          description: No content
          content:
            application/json:
              example:
                error: No content
        '404':
          description: Book not found
          content:
            application/json:
              example:
                error: Book not found
      tags:
        - Books